+++ src/test/java/com/atlassian/braid/SchemaBraidConsumerTest.java
@@ -5,8 +5,10 @@ import com.google.common.collect.ImmutableMap;
 import graphql.ExecutionInput;
 import graphql.ExecutionResult;
 import graphql.GraphQL;
+import graphql.execution.DataFetcherResult;
 import graphql.execution.instrumentation.dataloader.DataLoaderDispatcherInstrumentation;
 import graphql.schema.idl.TypeDefinitionRegistry;
+import org.dataloader.BatchLoader;
 import org.dataloader.DataLoaderRegistry;
 import org.junit.Rule;
 import org.junit.Test;
@@ -15,8 +17,8 @@ import org.mockito.Mock;
 import org.mockito.junit.MockitoJUnit;
 import org.mockito.junit.MockitoRule;
 
-import java.net.MalformedURLException;
 import java.util.Map;
+import java.util.concurrent.CompletableFuture;
 import java.util.function.Function;
 
 import static com.atlassian.braid.Util.parseRegistry;
@@ -24,14 +26,19 @@ import static graphql.ExecutionInput.newExecutionInput;
 import static graphql.schema.idl.RuntimeWiring.newRuntimeWiring;
 import static java.util.Collections.emptyList;
 import static java.util.Collections.emptyMap;
+import static java.util.Collections.singletonList;
 import static java.util.Collections.singletonMap;
 import static org.junit.Assert.assertEquals;
+import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Matchers.argThat;
+import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
+import static org.mockito.Mockito.withSettings;
 
 
+@SuppressWarnings("unchecked")
 public class SchemaBraidConsumerTest {
 
     private static final SchemaNamespace FOO = SchemaNamespace.of("foo");
@@ -43,7 +50,7 @@ public class SchemaBraidConsumerTest {
     private Function<ExecutionInput, Object> queryExecutor;
 
     @Test
-    public void testBraidWithExistingTypes() throws MalformedURLException {
+    public void testBraidWithExistingTypes() {
 
 
         final TypeDefinitionRegistry existingRegistry = parseRegistry("/com/atlassian/braid/existing.graphql");
@@ -83,6 +90,48 @@ public class SchemaBraidConsumerTest {
         final ExecutionResult result = graphql.execute(newExecutionInput().query(query).context(context));
 
         verify(queryExecutor, times(1)).apply(argThat(matchesInput(fooInput)));
+        assertEquals(emptyList(), result.getErrors());
+
+        Map<String, Map<String, Object>> data = result.getData();
+        assertEquals(data.get("foo").get("name"), "Foo");
+    }
+
+    @Test
+    public void testBraidWithLocalTypesAsSchemaSource() {
+
+
+        final TypeDefinitionRegistry existingRegistry = parseRegistry("/com/atlassian/braid/existing.graphql");
+        final TypeDefinitionRegistry fooRegistry = parseRegistry("/com/atlassian/braid/foo.graphql");
+
+        SchemaSource localSource = mock(SchemaSource.class, withSettings()
+                .extraInterfaces(BatchLoaderFactory.class));
+        when(localSource.getNamespace()).thenReturn(FOO);
+        when(localSource.getSchema()).thenReturn(fooRegistry);
+        BatchLoader loader = mock(BatchLoader.class);
+        when(loader.load(any())).thenReturn(CompletableFuture.completedFuture(
+                singletonList(new DataFetcherResult(
+                        ImmutableMap.of("id", "fooid", "name", "Foo"),
+                        emptyList()))));
+        when(((BatchLoaderFactory)localSource).newBatchLoader(any(), any())).thenReturn(loader);
+
+        Braid braid = new SchemaBraid<>()
+                .braid(SchemaBraidConfiguration.builder()
+                        .typeDefinitionRegistry(existingRegistry)
+                        .runtimeWiringBuilder(newRuntimeWiring())
+                        .schemaSource(localSource)
+                        .build());
+
+        DataLoaderRegistry dataLoaderRegistry = braid.newDataLoaderRegistry();
+
+        GraphQL graphql = new GraphQL.Builder(braid.getSchema())
+                .instrumentation(new DataLoaderDispatcherInstrumentation(dataLoaderRegistry))
+                .build();
+
+        String query = "{ foo(id: \"fooid\") { id, name } }";
+
+        final BraidContext context = new DefaultBraidContext(dataLoaderRegistry, emptyMap(), query);
+        final ExecutionResult result = graphql.execute(newExecutionInput().query(query).context(context));
+
         assertEquals(emptyList(), result.getErrors());
 
         Map<String, Map<String, Object>> data = result.getData();